Also moved ComputerPickle and FilePathPickle to workflow-durable-task-step, introducing FilePathUtils pending JENKINS-26096.
